clean tau, egamma, mu and hadron input functions
